将功率修正为浮点数，修改相关方法

+++ SES-pojo/src/main/java/com/SES/entity/DeviceLog.java
@@ -38,7 +38,7 @@ public class DeviceLog implements Serializable {
 
     private String policy; // JSON 格式，冗余信息
 
-    private Integer power; // 单位：W
+    private Float power; // 单位：W
 
     private Integer energyConsumption; // 单位：kWh
 }
